+++ src/edit.h
@@ -1,6 +1,15 @@
 #ifndef NEOVIM_EDIT_H
 #define NEOVIM_EDIT_H
-/* edit.c */
+
+/*
+ * Array indexes used for cptext argument of ins_compl_add().
+ */
+#define CPT_ABBR    0   /* "abbr" */
+#define CPT_MENU    1   /* "menu" */
+#define CPT_KIND    2   /* "kind" */
+#define CPT_INFO    3   /* "info" */
+#define CPT_COUNT   4   /* Number of entries */
+
 int edit(int cmdchar, int startln, long count);
 void edit_putchar(int c, int highlight);
 void edit_unputchar(void);
