+++ src/tag.h
@@ -1,6 +1,17 @@
 #ifndef NEOVIM_TAG_H
 #define NEOVIM_TAG_H
-/* tag.c */
+
+/*
+ * Structure used for get_tagfname().
+ */
+typedef struct {
+  char_u      *tn_tags;         /* value of 'tags' when starting */
+  char_u      *tn_np;           /* current position in tn_tags */
+  int tn_did_filefind_init;
+  int tn_hf_idx;
+  void        *tn_search_ctx;
+} tagname_T;
+
 int do_tag(char_u *tag, int type, int count, int forceit, int verbose);
 void tag_freematch(void);
 void do_tags(exarg_T *eap);
